Use Table component
